lbmap: Move type conversion routines to types.go
Again, to keep the module size sane.

No functional changes.
