+++ pkg/maps/lbmap/types.go
@@ -20,6 +20,10 @@ import (
 	"unsafe"
 
 	"github.com/cilium/cilium/pkg/bpf"
+	"github.com/cilium/cilium/pkg/loadbalancer"
+	"github.com/cilium/cilium/pkg/logging/logfields"
+
+	"github.com/sirupsen/logrus"
 )
 
 // BackendAddrID is the type of a service endpoint's unique identifier which
@@ -255,4 +259,166 @@ func (s *RRSeqValue) GetValuePtr() unsafe.Pointer { return unsafe.Pointer(s) }
 
 func (s *RRSeqValue) String() string {
 	return fmt.Sprintf("count=%d idx=%v", s.Count, s.Idx)
+}
+
+// l3n4Addr2ServiceKey converts the given l3n4Addr to a ServiceKey with the slave ID
+// set to 0.
+func l3n4Addr2ServiceKey(l3n4Addr loadbalancer.L3n4AddrID) ServiceKey {
+	log.WithField(logfields.L3n4AddrID, l3n4Addr).Debug("converting L3n4Addr to ServiceKey")
+	if l3n4Addr.IsIPv6() {
+		return NewService6Key(l3n4Addr.IP, l3n4Addr.Port, 0)
+	}
+	return NewService4Key(l3n4Addr.IP, l3n4Addr.Port, 0)
+}
+
+// LBSVC2ServiceKeynValue transforms the SVC Cilium type into a bpf SVC type.
+func LBSVC2ServiceKeynValue(svc loadbalancer.LBSVC) (ServiceKey, []ServiceValue, error) {
+	log.WithFields(logrus.Fields{
+		"lbFrontend": svc.FE.String(),
+		"lbBackend":  svc.BES,
+	}).Debug("converting Cilium load-balancer service (frontend -> backend(s)) into BPF service")
+	fe := l3n4Addr2ServiceKey(svc.FE)
+
+	// Create a list of ServiceValues so we know everything is safe to put in the lb
+	// map
+	besValues := []ServiceValue{}
+	for _, be := range svc.BES {
+		beValue := fe.NewValue().(ServiceValue)
+		if err := beValue.SetAddress(be.IP); err != nil {
+			return nil, nil, err
+		}
+		beValue.SetPort(be.Port)
+		beValue.SetRevNat(int(svc.FE.ID))
+		beValue.SetWeight(be.Weight)
+
+		besValues = append(besValues, beValue)
+		log.WithFields(logrus.Fields{
+			"lbFrontend": fe,
+			"lbBackend":  beValue,
+		}).Debug("associating frontend -> backend")
+	}
+	log.WithFields(logrus.Fields{
+		"lbFrontend":        svc.FE.String(),
+		"lbBackend":         svc.BES,
+		logfields.ServiceID: fe,
+		logfields.Object:    logfields.Repr(besValues),
+	}).Debug("converted LBSVC (frontend -> backend(s)), to Service Key and Value")
+	return fe, besValues, nil
+}
+
+// L3n4Addr2RevNatKeynValue converts the given L3n4Addr to a RevNatKey and RevNatValue.
+func L3n4Addr2RevNatKeynValue(svcID loadbalancer.ServiceID, feL3n4Addr loadbalancer.L3n4Addr) (RevNatKey, RevNatValue) {
+	if feL3n4Addr.IsIPv6() {
+		return NewRevNat6Key(uint16(svcID)), NewRevNat6Value(feL3n4Addr.IP, feL3n4Addr.Port)
+	}
+	return NewRevNat4Key(uint16(svcID)), NewRevNat4Value(feL3n4Addr.IP, feL3n4Addr.Port)
+}
+
+// serviceKey2L3n4Addr converts the given svcKey to a L3n4Addr.
+func serviceKey2L3n4Addr(svcKey ServiceKey) *loadbalancer.L3n4Addr {
+	log.WithField(logfields.ServiceID, svcKey).Debug("creating L3n4Addr for ServiceKey")
+	var (
+		feIP   net.IP
+		fePort uint16
+	)
+	if svcKey.IsIPv6() {
+		svc6Key := svcKey.(*Service6Key)
+		feIP = svc6Key.Address.IP()
+		fePort = svc6Key.Port
+	} else {
+		svc4Key := svcKey.(*Service4Key)
+		feIP = svc4Key.Address.IP()
+		fePort = svc4Key.Port
+	}
+	return loadbalancer.NewL3n4Addr(loadbalancer.NONE, feIP, fePort)
+}
+
+// serviceKeynValue2FEnBE converts the given svcKey and svcValue to a frontend in the
+// form of L3n4AddrID and backend in the form of L3n4Addr.
+func serviceKeynValue2FEnBE(svcKey ServiceKey, svcValue ServiceValue) (*loadbalancer.L3n4AddrID, *loadbalancer.LBBackEnd) {
+	var (
+		beIP     net.IP
+		svcID    loadbalancer.ServiceID
+		bePort   uint16
+		beWeight uint16
+	)
+
+	log.WithFields(logrus.Fields{
+		logfields.ServiceID: svcKey,
+		logfields.Object:    logfields.Repr(svcValue),
+	}).Debug("converting ServiceKey and ServiceValue to frontend and backend")
+
+	if svcKey.IsIPv6() {
+		svc6Val := svcValue.(*Service6Value)
+		svcID = loadbalancer.ServiceID(svc6Val.RevNat)
+		beIP = svc6Val.Address.IP()
+		bePort = svc6Val.Port
+		beWeight = svc6Val.Weight
+	} else {
+		svc4Val := svcValue.(*Service4Value)
+		svcID = loadbalancer.ServiceID(svc4Val.RevNat)
+		beIP = svc4Val.Address.IP()
+		bePort = svc4Val.Port
+		beWeight = svc4Val.Weight
+	}
+
+	feL3n4Addr := serviceKey2L3n4Addr(svcKey)
+	beLBBackEnd := loadbalancer.NewLBBackEnd(loadbalancer.NONE, beIP, bePort, beWeight)
+
+	feL3n4AddrID := &loadbalancer.L3n4AddrID{
+		L3n4Addr: *feL3n4Addr,
+		ID:       svcID,
+	}
+
+	return feL3n4AddrID, beLBBackEnd
+}
+
+func serviceValue2L3n4Addr(svcVal ServiceValue) *loadbalancer.L3n4Addr {
+	var (
+		beIP   net.IP
+		bePort uint16
+	)
+	if svcVal.IsIPv6() {
+		svc6Val := svcVal.(*Service6Value)
+		beIP = svc6Val.Address.IP()
+		bePort = svc6Val.Port
+	} else {
+		svc4Val := svcVal.(*Service4Value)
+		beIP = svc4Val.Address.IP()
+		bePort = svc4Val.Port
+	}
+	return loadbalancer.NewL3n4Addr(loadbalancer.NONE, beIP, bePort)
+}
+
+// RevNat6Value2L3n4Addr converts the given RevNat6Value to a L3n4Addr.
+func revNat6Value2L3n4Addr(revNATV *RevNat6Value) *loadbalancer.L3n4Addr {
+	return loadbalancer.NewL3n4Addr(loadbalancer.NONE, revNATV.Address.IP(), revNATV.Port)
+}
+
+// revNat4Value2L3n4Addr converts the given RevNat4Value to a L3n4Addr.
+func revNat4Value2L3n4Addr(revNATV *RevNat4Value) *loadbalancer.L3n4Addr {
+	return loadbalancer.NewL3n4Addr(loadbalancer.NONE, revNATV.Address.IP(), revNATV.Port)
+}
+
+// revNatValue2L3n4AddrID converts the given RevNatKey and RevNatValue to a L3n4AddrID.
+func revNatValue2L3n4AddrID(revNATKey RevNatKey, revNATValue RevNatValue) *loadbalancer.L3n4AddrID {
+	var (
+		svcID loadbalancer.ServiceID
+		be    *loadbalancer.L3n4Addr
+	)
+	if revNATKey.IsIPv6() {
+		revNat6Key := revNATKey.(*RevNat6Key)
+		svcID = loadbalancer.ServiceID(revNat6Key.Key)
+
+		revNat6Value := revNATValue.(*RevNat6Value)
+		be = revNat6Value2L3n4Addr(revNat6Value)
+	} else {
+		revNat4Key := revNATKey.(*RevNat4Key)
+		svcID = loadbalancer.ServiceID(revNat4Key.Key)
+
+		revNat4Value := revNATValue.(*RevNat4Value)
+		be = revNat4Value2L3n4Addr(revNat4Value)
+	}
+
+	return &loadbalancer.L3n4AddrID{L3n4Addr: *be, ID: svcID}
 }
